Use fixed size ints in rpc structs
image.Image, image.Info and path.Thumbnail had int for width and height,
which breaks java.
Also forced image.Info to still be called ImageInfo in java.

+++ gfxapi/texture.go
@@ -30,18 +30,18 @@ type Texture struct {
 }
 
 // Thumbnail returns the image that most closely matches the desired size.
-func (t *Texture) Thumbnail(w, h int, d database.Database, l log.Logger) (*image.Info, error) {
+func (t *Texture) Thumbnail(w, h uint32, d database.Database, l log.Logger) (*image.Info, error) {
 	if len(t.Levels) == 0 {
 		return nil, errors.New("Texture has no levels")
 	}
 
 	var best *image.Info
-	bestScore := 0xffffffff
+	bestScore := uint32(0xffffffff)
 
-	sqr := func(i int) int { return i * i }
+	sqr := func(i uint32) uint32 { return i * i }
 
 	for _, level := range t.Levels {
-		score := sqr(int(level.Width)-w) + sqr(int(level.Height)-h)
+		score := sqr(level.Width-w) + sqr(level.Height-h)
 		if bestScore > score {
 			level := level
 			best, bestScore = &level, score
